requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
cairocffi==0.9.0
CairoSVG==2.2.1
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.5
django-bootstrap==0.2.4
django-celery==3.2.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-heroku==0.3.1
django-paypal==0.5.0
django-rest-framework==0.1.0
django-summernote==0.8.11.4
django-taggit==0.23.0
django-tinymce==2.8.0
djangorestframework==3.9.0
html5lib==1.0.1
idna==2.8
kombu==3.0.37
Markdown==3.0.1
Pillow==5.4.0
ply==3.11
psycopg2==2.7.6.1
pycparser==2.19
Pyphen==0.9.5
python-dateutil==1.5
pytz==2018.7
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tinycss2==0.6.1
tqdm==4.28.1
urllib3==1.24.1
virtualenv==16.2.0
virtualenvwrapper-win==1.2.5
WeasyPrint==44
webencodings==0.5.1
whitenoise==4.1.2
